Move scheduleAfter to JobUtils
